ci: Fix coverage report
Coverage reporting on the package installed in the virtualenv had the
undesired side effect of listing the filenames in the virtualenv instead
of the "local" files. Rolls back to editable install and the path name
for `py.test --cov`.

Reverts 147186b130efb57032dbce9c8a88d37447319aa6 and part of
31c8b3e152745e9b90e348a1ebd24c78dd54b07b.