Ionization by secondary electrons via bremsstrahlung
Implements the coefficients as listed in Berger et al., 1974 for
ionization by secondary electrons from bremsstrahlung.

The function signature is similar to the electron ionization functions,
and adding the shape test for it was trivial.

The units are still a bit off, but the profiles for exponential primary
electrons look very close to what is shown in their Fig. 7.