electrons: Use more sophisticated `polyval`
The `polyval` function in `numpy.polynomial.polynomial` is much more
useful for this purpose. Unfortunately it is well-hidden in the docs,
someone could have told us much earlier about it.

Splits out the Fang polynomial-exponential evaluation since it is the
same for Fang2008 and Fang2010, removing duplicate code.
Note that since we use a different `polyval` function, there is no need
to reshape the evaluated coefficient matrix for that.