e-: Fix multidimensional Fang2010 integration
We cannot rely on a certain order of the dimensions of the
energy/fluxes versus the rho/scale_height arrays.
Numpy takes care of that in all the other functions, but the Fang2010 spectral
integration needs the spectrum array in a certain order.

Fixes the Fang2010 Maxwellian spectral shape for transposed
(and higher dimensional) arrays.