tests: Add transposed and 3d test cases
Ensures that the shapes are correct and the expected values are in the
right places when the input arrays have different shapes and are of
higher dimensions.

Exclude bremsstrahlung for now, scipy's rbf interpolation uses
np.meshgrid which messes with the order of the dimensions and doesn't
(yet) work for correctly for higher-dimensional arrays.