F2010: Fix scalar input and add tests
The integration over the Fang2010 mono-energetic spectrum uses `.ndim`
to construct the shape of the spectrum array, which is not available if
scalars are passed to the functions. Converts the inputs to the
to `numpy.ndarray`s to have `.ndim` available.

Adds a test that scalar inputs for energies and fluxes work as expected.