brems: Fix output shape with `Rbf()`
Using `np.meshgrid()` resulted in transposed or non-complete energy/z
arrays for the interpolation with `scipy`'s radial basis functions.

Reshapes the input energy and z arrays to have the same (automatically
broadcasted) shape as two columns for the to-be interpolated array of
2-D points. This prevents `Rbf` from complaining about un-euqal array
lengths and automatically gives the right shape of the bremsstrahlung
coefficients.

Includes the bremsstrahlung parametrization in the different shape
tests, they were excluded because they failed and should now pass.