More spectra variants
Adds the particle-flux and energy-flux normalized spectra of
exponential, Gaussian, and power-law type, in addition to the already
available Maxwellian type spectra.

The particle-flux normalized versions are basically standard probability
distributions. We assume that the negative tail for the Gaussian
distribution does not contribute much and can be neglected.

The energy-flux normalized variants `pflux_...()` should be useful as
they are, scaled by the integrated energy flux.