spectra: Enable scalar input for power-law fluxes
We select the zero spots by directly accessing the elements, which fails
if the energy (`en`) is a scalar. Temporarily converts the input to a
(at least) 1-d array so that indexing works. Converts it back if a
scalar was passed such that input and output have the same shape.

This is only necessary for the power law type functions, where we
actually rely on the indexing to work. The other functions take `numpy`
arrays which keep their shape intact.