ci: Upload coverage report to coveralls
Uploads the code coverage reports also to coveralls, which is a bit more
complicated than the `codecov` action.

Runs coverage in parallel mode, needed to produce a per-host .coverage
file to be combined later.
Reference:
https://coverage.readthedocs.io/en/6.2/cmd.html#combining-data-files-coverage-combine

Using `pytest`s coverage module calls `combine` to produce the report
which fails in this approach.
Uses `coverage`s `--source` option and a separate `report` call to get
the report instead of using `pytest-cov` for that.

Excludes python 2.7 from coveralls upload, as the `coveralls` package
seems to be too old.