models: Rewrite "offset" name detection and match
Depending on the python/xarray combination, the names of proxies in the
coefficient file are read as strings or bytes. When the latter is the
case, a simple match against "offset" fails, resulting in a false
negative, so convert the name to bytes in that case.
Should fix the Python 3.6 tests on github CI and should be more robust.