models: Warn when swapping proxy axis
Be a bit more verbose when re-ordering the proxy array based on the size
of the dimensions. Takes care that the zero-th axis (rows) matches the
proxies in the coefficient table. Warns the user that this is ambiguous
in cases where there is more than one such axis that matches this dimension.
In that case the first one is taken.