models: Update SSUSI model coefficients
Updates the SSUSI ionization rate model coefficients to the ones derived
from the "h" version fits in the Northern hemisphere, and "g" in the
Southern hemisphere.

"h" (NH) - Combined data files with float64 and float32 data which
introduced float representation errors in the julian date which was used
to find the proxy times. Fixed by rounding the julian date to 4 digits
compared to the "g" version.

"g" (SH) - Data correctly taken from the SH compared to "d".