Update SSUSI model coefficients to v2
Removes the "log_v_plasma" proxy from the model coefficients because
it interfered too much with the offset, especially in cells with low
number of data points.

Uses bi-linear interpolation in MLT/latitude to fill cells with low
number of data points. Tries to be verbose about that in the data file.

Reduces all parameters to single precision which is sufficient and saves
some space, in particular for distributing the package.