models: Fix proxy selection and dtype
With Py36 there is a mismatch between byte and string dtypes so that
the down-selection of common proxies fails. Converts the proxy names to
the same dtype so that `xarray` manages to select the right ones.

Also shifts the selection one ident level, we want to select every time,
not only when there is "offset" in the proxy list.