models: Helper function for `xarray` interpolation
Uses a helper function to interpolate the coefficient data set.
`xarray`'s own `.interp()` does not work when the data set contains
non-numeric variables, so we try to re-create the behaviour that it is
supposed to have according to the documentation.
Numeric variables are interpolated, non-numeric ones `.sel()`ected, the
method for non-numeric types is currently fixed but should be good
enough for the purpose, all we need is `beta` and `beta_std`.